# docs/frontleaves_plugin_swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/servers/game-profiles/online/mine": {
            "get": {
                "description": "查询当前登录用户所有在线的游戏账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "玩家在线查询接口"
                ],
                "summary": "[用户] 获取在线游戏账号",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/apiServer.OnlineGameProfileResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/servers/players/online/check": {
            "get": {
                "description": "按 UUID 或用户名检查指定玩家是否在线",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "玩家在线查询接口"
                ],
                "summary": "[用户] 检查玩家在线状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家 UUID",
                        "name": "uuid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "玩家用户名",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apiServer.PlayerOnlineResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apiServer.OnlineGameProfileResponse": {
            "type": "object",
            "properties": {
                "server_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "world_name": {
                    "type": "string"
                }
            }
        },
        "apiServer.PlayerOnlineResponse": {
            "type": "object",
            "properties": {
                "last_seen": {
                    "type": "integer"
                },
                "online": {
                    "type": "boolean"
                },
                "player_name": {
                    "type": "string"
                },
                "player_uuid": {
                    "type": "string"
                },
                "server_name": {
                    "type": "string"
                },
                "world_name": {
                    "type": "string"
                }
            }
        },
...
        "xBase.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "string"
                },
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "overhead": {
                    "type": "integer"
                }
            }
        }
    }
}